MarX Version 1.0
*** About the MarX Application*

This application prints provider and carrier data retrieved from the
iPhone, after the “show log” button being pressed.    It can be used
for debugging and handling iPhone problems along with the plug log.
the application will continue printing data even when in the background
or while phone is in sleep mode. It does this by playing a silent music
file infinitely.

*** Version Features **

1. Print MCC, MNC, Carrier name, Signal Strength, SIM Status and Data
connectivity every 1 second.

2. Provides “Show Log” and “Stop Log” options.

Known Issus:

1. If the phone is turned off, the log is deleted, because the cache
files on iOS are removed.

2. Currently the app does not print the log in the application window.

Future Versions:

1. Support keeping log file after phone reboot.

2. Support printing text on the app window

3. Get Additional info, such as APN and etc.